# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blog",
        "description": "Documentación de la API (Laravel 11 + Passport).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor base"
        }
    ],
    "paths": {
...
        "/api/cards": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Obtener todas las tarjetas (Cards)",
                "description": "Devuelve una lista de todas las tarjetas con sus detalles de categoría y método de comunicación.",
                "operationId": "0cf776c75245f5f70e78d1981e3398c2",
                "responses": {
                    "200": {
                        "description": "Lista de tarjetas obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CardResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cards"
                ],
                "summary": "Crear una nueva tarjeta",
                "operationId": "6a1db83adb51fba2ac318a78bc2a4fde",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarjeta creada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "No autorizado (Requiere rol therapist/admin)."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cards/{card_id}": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Obtener una tarjeta por ID",
                "description": "Devuelve los detalles de una tarjeta específica por su clave primaria.",
                "operationId": "7350b288de2e917f6564caa984f7f7db",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarjeta obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarjeta no encontrada."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cards"
                ],
                "summary": "Eliminar una tarjeta",
                "operationId": "46d519107786a6310019d149617cdfa0",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarjeta eliminada exitosamente (No Content)."
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "No autorizado (Requiere rol therapist/admin)."
                    },
                    "404": {
                        "description": "Tarjeta no encontrada o no pudo ser eliminada."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Cards"
                ],
                "summary": "Actualizar una tarjeta existente",
                "operationId": "2031d580adc9422592399021c588e1de",
                "parameters": [
                    {
                        "name": "card_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCardRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarjeta actualizada exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado."
                    },
                    "403": {
                        "description": "No autorizado (Requiere rol therapist/admin)."
                    },
                    "404": {
                        "description": "Tarjeta no encontrada."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cards/uuid/{uuid}": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Obtener una tarjeta por UUID",
                "description": "Devuelve los detalles de una tarjeta específica por su UUID único.",
                "operationId": "cfa9fee4343802c9da342309117bfee4",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarjeta obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CardResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarjeta no encontrada."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCardRequest": {
                "title": "Store Card Request",
                "description": "Datos necesarios para crear una nueva Card. El 'uuid' es opcional y se genera si no se proporciona.",
                "required": [
                    "imagePath",
                    "methodId",
                    "categoryId"
                ],
                "properties": {
                    "uuid": {
                        "description": "UUID único de la tarjeta (se genera automáticamente si no se envía).",
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                        "nullable": true
                    },
                    "imagePath": {
                        "description": "Ruta o URL de la imagen de la tarjeta.",
                        "type": "string",
                        "example": "images/nueva_card.png"
                    },
                    "methodId": {
                        "description": "ID del método de comunicación.",
                        "type": "integer",
                        "example": 1
                    },
                    "categoryId": {
                        "description": "ID de la categoría.",
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "UpdateCardRequest": {
                "title": "Update Card Request",
                "description": "Datos opcionales para actualizar una Card existente. Utiliza la lógica 'sometimes' para permitir actualizaciones parciales (PATCH).",
                "properties": {
                    "uuid": {
                        "description": "UUID único de la tarjeta. Solo se valida si se envía.",
                        "type": "string",
                        "format": "uuid",
                        "example": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                        "nullable": true
                    },
                    "imagePath": {
                        "description": "Ruta o URL de la imagen de la tarjeta.",
                        "type": "string",
                        "example": "images/nueva_card_v2.png",
                        "nullable": true
                    },
                    "methodId": {
                        "description": "ID del método de comunicación.",
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "categoryId": {
                        "description": "ID de la categoría.",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "CardResource": {
                "title": "Card Resource",
                "description": "Representación de una tarjeta de comunicación (Card) con sus datos de identificación y relaciones.",
                "properties": {
                    "cardId": {
                        "description": "ID primario de la tarjeta.",
                        "type": "integer",
                        "example": 52
                    },
                    "uuid": {
                        "description": "UUID único de la tarjeta.",
                        "type": "string",
                        "format": "uuid",
                        "example": "8a0e0a9e-9c19-4d50-9e91-54dae0705053"
                    },
                    "imagePath": {
                        "description": "Ruta o URL de la imagen.",
                        "type": "string",
                        "example": "placeholders/card_image_52.jpg"
                    },
                    "methodId": {
                        "description": "ID del método de comunicación.",
                        "type": "integer",
                        "example": 2
                    },
                    "categoryIdCard": {
                        "description": "ID de la categoría.",
                        "type": "integer",
                        "example": 2
                    },
                    "categoryName": {
                        "description": "Nombre de la categoría.",
                        "type": "string",
                        "example": "Sentimientos",
                        "nullable": true
                    },
                    "methodName": {
                        "description": "Nombre del método de comunicación.",
                        "type": "string",
                        "example": "Auditivo",
                        "nullable": true
                    },
                    "consecutiveId": {
                        "description": "ID consecutivo/índice de ordenamiento. Solo presente en listados (index).",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}